split out ic_principal
