+++ Cargo.toml
@@ -3,6 +3,7 @@ members = [
     "rust/candid",
     "rust/candid_parser",
     "rust/candid_derive",
+    "rust/ic_principal",
     "tools/didc",
 ]
 resolver = "2"
