+++ rust/candid/Cargo.toml
@@ -19,9 +19,8 @@ include = ["src", "Cargo.toml", "build.rs", "LICENSE", "README.md"]
 byteorder = "1.4.3"
 candid_derive = { path = "../candid_derive", version = "=0.6.3" }
 codespan-reporting = "0.11"
-crc32fast = "1.3.0"
-data-encoding = "2.4.0"
 hex = "0.4.2"
+ic_principal = { path = "../ic_principal", version = "0.1.0" }
 leb128 = "0.2.4"
 num_enum = "0.6.1"
 num-bigint = { version = "0.4.2", features = ["serde"] }
@@ -30,7 +29,6 @@ paste = "1.0.0"
 pretty = "0.12.0"
 serde = { version = "1.0.118", features = ["derive"] }
 serde_bytes = "0.11"
-sha2 = "0.10.1"
 thiserror = "1.0.20"
 anyhow = "1.0"
 binread = { version = "2.1", features = ["debug_template"] }
@@ -43,8 +41,6 @@ rand = "0.8"
 criterion = "0.4"
 serde_cbor = "0.11.2"
 serde_json = "1.0.74"
-serde_test = "1.0.137"
-impls = "1"
 bincode = "1.3.3"
 
 [[bench]]
