+++ rust/ic_principal/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "ic_principal"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+crc32fast = "1.3.0"
+data-encoding = "2.3.2"
+hex = "0.4.3"
+sha2 = "0.10.1"
+thiserror = "1.0.30"
+
+[dev-dependencies]
+serde = { version = "1.0.133", features = ["derive"] }
+serde_cbor = "0.11.2"
+serde_json = "1.0.74"
+serde_test = "1.0.137"
+impls = "1"
+
+[dependencies.serde]
+version = "1.0.115"
+features = ["derive"]
+optional = true
+
+[dependencies.serde_bytes]
+version = "0.11.5"
+optional = true
+
+[features]
+# Default features include serde support.
+default = ['serde', 'serde_bytes']
