+++ rust/ic_principal/tests/principal.rs
@@ -1,6 +1,7 @@
-#![allow(deprecated)]
+// #![allow(deprecated)]
 
-use candid::types::principal::{Principal, PrincipalError};
+use ic_principal::Principal;
+use ic_principal::PrincipalError;
 
 const MANAGEMENT_CANISTER_BYTES: [u8; 0] = [];
 const MANAGEMENT_CANISTER_TEXT: &str = "aaaaa-aa";
@@ -263,7 +264,7 @@ fn self_authenticating_ok() {
     // 1. Sha224 hash is used
     // 2. 0x02 was added in the end
     // 3. total length is 29
-    let p1 = Principal::self_authenticating([]);
+    let p1 = Principal::self_authenticating(&[]);
     let p2 = Principal::try_from_slice(&[
         209, 74, 2, 140, 42, 58, 43, 201, 71, 97, 2, 187, 40, 130, 52, 196, 21, 162, 176, 31, 130,
         142, 166, 42, 197, 179, 228, 47, 2,
